spider/xhsSpider/requirements.txt
PyExecJS
requests
loguru
python-dotenv
retry
openpyxl
pillow
numpy==1.26.4
paddlepaddle==2.6.2
paddleocr==2.7.0.3